simplify

+++ src/wasm/literal.cpp
@@ -1073,8 +1073,8 @@ Literal Literal::subSatUI16(const Literal& other) const {
 Literal Literal::q15MulrSatSI16(const Literal& other) const {
   int64_t value =
     (int64_t(geti32()) * int64_t(other.geti32()) + 0x4000LL) >> 15LL;
-  int64_t lower = int64_t(std::numeric_limits<int16_t>::min());
-  int64_t upper = int64_t(std::numeric_limits<int16_t>::max());
+  int64_t lower = std::numeric_limits<int16_t>::min();
+  int64_t upper = std::numeric_limits<int16_t>::max();
   return Literal(int16_t(std::min(std::max(value, lower), upper)));
 }
